fix(elements): globalized footer
