+++ pages/about.tsx
@@ -1,0 +1,8 @@
+
+export default function About() {
+    return (
+        <div className="container">
+            
+        </div>
+    );
+}
