+++ styles/globals.scss
@@ -2,7 +2,7 @@ html,
 body {
   padding: 0;
   margin: 0;
-  font-family: -apple-system, BlinkMacSystemFont, Segoe UI, Roboto, Oxygen,
+  font-family: Nexa, Raleway, Roboto, Oxygen,
     Ubuntu, Cantarell, Fira Sans, Droid Sans, Helvetica Neue, sans-serif;
 }
 
@@ -15,12 +15,56 @@ a {
   box-sizing: border-box;
 }
 
+.container {
+  padding: 0 2rem;
+  display: flex;
+  flex-direction: column;
+  height: 100%;
+}
+
+footer {
+  display: flex;
+  flex-wrap: wrap;
+  flex-direction: row;
+  flex: 1;
+  border-top: 3px solid #eaeaea;
+  padding: 1.6rem 0;
+  align-items: center;
+  justify-content: center;
+  position: absolute;
+  width: 100%;
+  font-weight: 500;
+}
+
+footer a {
+  display: flex;
+  justify-content: center;
+  flex-grow: 1;
+}
+
+.vlogo {
+  height: 1em;
+  margin-left: 0.6rem;
+}
+
+.logo {
+  margin-top: -0.8rem;
+  height: 1em;
+  margin-left: 0.6rem;
+}
+
 @media (prefers-color-scheme: dark) {
+  footer {
+    border-color: #222;
+  }
   html {
     color-scheme: dark;
   }
   body {
     color: white;
     background: black;
+  }
+  .vlogo img {
+    filter: invert(1);
   }
 }
